Replace if statements/while loops during building of AST and not after
