+++ src/lib/compiler/mod.rs
@@ -1,6 +1,7 @@
 pub mod compile;
-pub mod if_statements;
-pub(crate) mod macros;
 pub mod parse;
+
+pub(crate) mod if_while;
+pub(crate) mod macros;
 
 pub struct Compiler;
